refactoring : cleaning
