+++ src/app/providers/index.ts
@@ -1,4 +1,6 @@
 import { router } from "./router/router";
+import { firestoreMiddleware } from "./store/middlewares/firestoreMiddlware";
+import { AuthLogsMiddleware } from "./store/middlewares/loggerMiddleware";
 import { store } from "./store/store";
 
-export { router, store };
+export { router, store, AuthLogsMiddleware, firestoreMiddleware };
