+++ src/app/providers/store/store.tsx
@@ -5,8 +5,7 @@ import { favoriteReducer } from "../../../shared/reducers/Favorites";
 import { firestoreReducer } from "../../../shared/reducers/Profile/Firestore";
 import { searchReducer } from "../../../shared/reducers/Search";
 import { historyReducer } from "../../../shared/reducers/History/";
-import { AuthLogsMiddleware } from "./middlewares/loggerMiddleware";
-import { firestoreMiddleware } from "./middlewares/firestoreMiddlware";
+import { AuthLogsMiddleware, firestoreMiddleware } from "../";
 
 export const store = configureStore({
   reducer: {
