+++ src/pages/SearchPage/ui/SearchPage.tsx
@@ -1,6 +1,6 @@
 import { useSearchParams } from "react-router-dom";
 import { useSearch } from "../../../shared/hooks/useSearch";
-import { SearchItems } from "../../../widgest/Search";
+import { SearchItems } from "../../../widgest/SearchItems";
 import { useEffect } from "react";
 import { setResultSearch } from "../../../shared/reducers/Search";
 import { AppDispatch } from "../../../app/providers/store/store";
