+++ src/shared/reducers/Authorization/slice/Authslice.ts
@@ -41,7 +41,7 @@ const authSlice = createSlice({
         state.isAuthenticated = true;
       })
 
-      //Не смог их типизировать но хотел чтобы показывались именно ошибки FIREBASE Могу переписать на обычные ошибки
+      //Не смог их типизировать но хотел чтобы показывались именно ошибки FIREBASE Могу переписать на обычные ошибки без any
       //без этого выдавал ошибку
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       .addCase(signup.rejected, (state, action: any) => {
