+++ src/shared/reducers/Favorites/index.ts
@@ -2,6 +2,18 @@ import {
   addToFavorites,
   deleteFromFavorites,
 } from "./actions/FavoritesActions";
+import {
+  isFirstLoadingFavorites,
+  getErrorFavorites,
+  getFavorites,
+} from "./selectors/FavoriteSelectors";
 import favoriteReducer from "./slice/FavoriteSlice";
 
-export { deleteFromFavorites, addToFavorites, favoriteReducer };
+export {
+  deleteFromFavorites,
+  addToFavorites,
+  isFirstLoadingFavorites,
+  getErrorFavorites,
+  getFavorites,
+  favoriteReducer,
+};
