+++ src/shared/reducers/Search/SearchSelectors/SearchSelector.ts
@@ -5,5 +5,3 @@ export const getSearchMovies = (state: RootState) => state.search.movies;
 export const getSearchStatus = (state: RootState) => state.search.status;
 
 export const getSearchError = (state: RootState) => state.search.isError;
-
-export const getSearchKeyword = (state: RootState) => state.search.keyword;
