+++ src/shared/reducers/Search/SearchSlice/SearchSlice.tsx
@@ -5,14 +5,12 @@ interface searchState {
   movies: Item[];
   status: "uninitialized" | "pending" | "fulfilled" | "rejected" | undefined;
   isError: boolean;
-  keyword: string | undefined;
 }
 
 const initialState: searchState = {
   movies: [],
   status: undefined,
   isError: false,
-  keyword: undefined,
 };
 
 const searchSlice = createSlice({
@@ -23,7 +21,6 @@ const searchSlice = createSlice({
       state.movies = action.payload.movies;
       state.isError = action.payload.isError;
       state.status = action.payload.status;
-      state.keyword = action.payload.keyword;
     },
   },
 });
