+++ src/shared/reducers/Search/index.ts
@@ -1,8 +1,15 @@
 import {
-  getSearchError,
+  getSearchStatus,
   getSearchMovies,
+  getSearchError,
 } from "./SearchSelectors/SearchSelector";
 import { setResultSearch } from "./SearchSlice/SearchSlice";
 import searchReducer from "./SearchSlice/SearchSlice";
 
-export { setResultSearch, searchReducer, getSearchMovies, getSearchError };
+export {
+  setResultSearch,
+  getSearchStatus,
+  searchReducer,
+  getSearchMovies,
+  getSearchError,
+};
