--- src/widgest/Loading/index.ts
@@ -1,3 +1,0 @@
-import { Loading } from "./ui/Loading";
-
-export { Loading };
