--- src/widgest/Loading/ui/Loading.tsx
@@ -1,3 +1,0 @@
-export function Loading() {
-  return <h2>Loading...</h2>;
-}
